requirements/project.txt
Mezzanine==1.3.0
django-classy-tags
django-extensions
mezzanine-slides
south
django-compressor
django-debug-toolbar
